+++ src/renderer/src/controller/index.ts
@@ -1,0 +1,1 @@
+export {  viewController } from './viewController'
