+++ src/renderer/src/layouts/index.ts
@@ -1,0 +1,1 @@
+export { default as HoverLayout } from './HoverLayout'
